requirements.txt
flask
python-dotenv
euriai